# src/main/resources/static/openapi/product-service-openapi.yaml
openapi: 3.0.1
info:
  title: Bank Product API
  version: 1.0.0
  description: API para gestionar productos bancarios (cuentas, créditos, tarjetas)
servers:
  - url: http://localhost:8081/product

paths:
  /debit-cards:
    post:
      summary: Registrar una nueva tarjeta de débito
      operationId: createDebitCard
      tags:
        - DebitCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardBody'
      responses:
        '200':
          description: Tarjeta registrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardTemplateResponse'

  /debit-cards/{id}:
    get:
      summary: Obtener tarjeta de débito por ID
      operationId: getDebitCardById
      tags:
        - DebitCards
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Identificador único de la tarjeta
      responses:
        '200':
          description: Tarjeta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardTemplateResponse'
        '404':
          description: Tarjeta no encontrada

  /products/customer/{id}:
    get:
      summary: Obtener un producto por su codigo de cliente
      description: Busca un producto bancario usando su codigo de cliente
      operationId: getAllProductsByCustomerId
      parameters:
        - name: id
          in: path
          required: true
          description: ID del cliente del producto a consultar
          schema:
            type: string
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankProductTemplateResponse'

  /products:
    get:
      summary: Obtener todos los productos
      description: Devuelve la lista completa de productos bancarios
      operationId: getAllProducts
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankProductTemplateResponse'

    post:
      summary: Crear un nuevo producto bancario
      description: Registra un nuevo producto (cuenta, crédito, tarjeta) asociado a un cliente
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankProductBody'
      responses:
        '200':
          description: Producto creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankProductTemplateResponse'

  /products/{id}:
    get:
      summary: Obtener un producto por ID
      description: Busca un producto bancario usando su ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: ID del producto a consultar
          schema:
            type: string
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankProductTemplateResponse'

    put:
      summary: Actualizar un producto bancario
      description: Actualiza los datos de un producto existente mediante su ID
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          description: ID del producto a actualizar
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankProductBody'
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankProductTemplateResponse'

    delete:
      summary: Eliminar un producto bancario
      description: Elimina un producto por su ID
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          description: ID del producto a eliminar
          schema:
            type: string
      responses:
        '200':
          description: Producto eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankProductTemplateResponse'

components:
  schemas:
    BankProductBody:
      type: object
      properties:
        customerId:
          type: string
          description: ID del cliente asociado al producto
          example: cust1234
        type:
          type: string
          description: Tipo de producto bancario
          example: "[ SAVINGS, CURRENT, FIXED_TERM, CREDIT ]"
        status:
          type: string
          description: Estado de producto bancario
          example: "[ ACTIVE, BLOCKED_AVG_BALANCE ]"
        name:
          type: string
          description: Nombre descriptivo del producto
          example: Cuenta Ahorro Soles
        balance:
          type: number
          format: decimal
          description: Saldo actual del producto
          example: 1000.50
        holders:
          type: array
          description: Titulares de la cuenta empresarial
          items:
            type: string
          example: [ "12345678", "87654321" ]
        signers:
          type: array
          description: Firmantes autorizados
          items:
            type: string
          example: [ "99887766" ]
        details:
          type: object
          description: Detalles del producto como JSON genérico
          example:
            maintenanceFee: 0
            monthlyLimit: 2
            requiredMonthlyAverageBalance: 1000

    CardBody:
      type: object
      description: Tarjeta de débito asociada a múltiples cuentas bancarias
      properties:
        cardNumber:
          type: string
          example: "1234-5678-9012-3456"
        customerId:
          type: string
          example: cust001
        primaryAccountId:
          type: string
          example: acc001
        linkedAccountIds:
          type: array
          items:
            type: string
          example: [ "acc001", "acc002", "acc003" ]
        active:
          type: boolean
          example: true

    CardResponse:
      type: object
      description: Tarjeta de débito asociada a múltiples cuentas bancarias
      properties:
        id:
          type: string
          example: card-001
        cardNumber:
          type: string
          example: "1234-5678-9012-3456"
        customerId:
          type: string
          example: cust001
        primaryAccountId:
          type: string
          example: acc001
        linkedAccountIds:
          type: array
          items:
            type: string
          example: [ "acc001", "acc002", "acc003" ]
        active:
          type: boolean
          example: true

    BankProductResponse:
      type: object
      properties:
        id:
          type: string
          description: ID del producto
          example: id1234
        customerId:
          type: string
          description: ID del cliente asociado al producto
          example: cust1234
        type:
          type: string
          description: Tipo de producto bancario
          example: "[ SAVINGS, CURRENT, FIXED_TERM, CREDIT ]"
        status:
          type: string
          description: Estado de producto bancario
          example: "[ ACTIVE, BLOCKED_AVG_BALANCE ]"
        name:
          type: string
          description: Nombre descriptivo del producto
          example: Cuenta Ahorro Soles
        balance:
          type: number
          format: decimal
          description: Saldo actual del producto
          example: 1000.50
        holders:
          type: array
          description: Titulares de la cuenta empresarial
          items:
            type: string
          example: [ "12345678", "87654321" ]
        signers:
          type: array
          description: Firmantes autorizados
          items:
            type: string
          example: [ "99887766" ]
        details:
          type: object
          description: Detalles del producto como JSON genérico
          example:
            maintenanceFee: 0
            monthlyLimit: 2

    BankProductTemplateResponse:
      type: object
      description: Respuesta procesada
      properties:
        status:
          type: integer
          description: Código HTTP de respuesta
          example: 200
        message:
          type: string
          description: Mensaje asociado a la operación
          example: Producto procesado correctamente
        products:
          type: array
          description: Lista de productos bancarios
          items:
            $ref: '#/components/schemas/BankProductResponse'

    CardTemplateResponse:
      type: object
      description: Respuesta procesada
      properties:
        status:
          type: integer
          description: Código HTTP de respuesta
          example: 200
        message:
          type: string
          description: Mensaje asociado a la operación
          example: Producto procesado correctamente
        cards:
          type: array
          description: Lista de productos bancarios
          items:
            $ref: '#/components/schemas/CardResponse'